update the readme, run out of road
This is as far as the 'write you a scheme' takes me. A regression was introduced in the last few commits. A test suite will help. Compliance tests will also be interesting, though I don't know that canonical spec tests exist. Given the size of the language features that remain I'm not sure when I'll get to them. Maybe I'll stick to rs4 intead. It could stand some refactoring to break out the parser and other components.